Added bulk of the fetching procedure
